feat: add query suggestions; refactor and lift up state/hooks to parent SearchPanel component
