+++ components/LoadingSpinner.tsx
@@ -1,0 +1,19 @@
+import React from "react";
+import { StyleSheet, ActivityIndicator, View } from "react-native";
+
+const LoadingSpinner = () => {
+  return (
+    <View style={styles.loadingContainer}>
+      <ActivityIndicator size="large" color="#FFFFFF" />
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  loadingContainer: {
+    flex: 1,
+    justifyContent: "center",
+  },
+});
+
+export { LoadingSpinner };
