+++ components/index.ts
@@ -1,7 +1,4 @@
-export { InfiniteHits } from "./InfiniteHits";
-export { SearchBox } from "./SearchBox";
-export { Highlight } from "./Highlight";
-export { Hit } from "./Hit";
-export { ListEmpty } from "./ListEmpty";
+export { SearchPanel } from "./search/SearchPanel";
 export { Typography, TextVariant } from "./Typography";
 export { Button, ButtonVariant } from "./Button";
+export { LoadingSpinner } from "./LoadingSpinner";
